requirements.txt
appdirs==1.4.3
Django==1.11.1
django-bootstrap3==8.2.3
django-jquery==3.1.0
django-treebeard==4.1.0
django-treemenus==0.9.2
packaging==16.8
psycopg2==2.7.1
pyparsing==2.2.0
pytz==2017.2
six==1.10.0
